requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
lxml==5.1.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-docx==1.1.0
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
